requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
mysql-connector==2.1.6
hurry.filesize==0.9

